Move atlas/util/mdspan.h to atlas/mdspan.h
