+++ src/atlas/grid/detail/grid/Unstructured.h
@@ -20,8 +20,8 @@
 #include <memory>
 #include <vector>
 
-#include "atlas/util/mdspan.h"
 #include "atlas/grid/detail/grid/Grid.h"
+#include "atlas/mdspan.h"
 #include "atlas/runtime/Exception.h"
 #include "atlas/util/Point.h"
 
@@ -153,7 +153,7 @@ public:  // methods
     /// Constructor taking a mdspan (makes copy)
     /// First dimension is number of points, second dimension is coordinate. First X (lon), then Y (lat)
     template <typename Extents, typename LayoutPolicy, typename AccessorPolicy>
-    Unstructured(atlas::mdspan<const double, Extents, LayoutPolicy, AccessorPolicy> xy):
+    Unstructured(mdspan<const double, Extents, LayoutPolicy, AccessorPolicy> xy):
         Unstructured(xy.extent(0), &xy[std::array{0,0}], &xy[std::array{0,1}], xy.stride(1), xy.stride(1)) {}
 
     /// Constructor from initializer list
