+++ src/atlas/mdspan.h
@@ -1,0 +1,25 @@
+/*
+ * (C) Copyright 2025- ECMWF.
+ *
+ * This software is licensed under the terms of the Apache Licence Version 2.0
+ * which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+ * In applying this licence, ECMWF does not waive the privileges and immunities
+ * granted to it by virtue of its status as an intergovernmental organisation
+ * nor does it submit to any jurisdiction.
+ */
+
+#pragma once
+
+#include "pluto/mdspan.h"
+
+namespace atlas {
+using ::pluto::dynamic_extent;
+using ::pluto::layout_left;
+using ::pluto::layout_right;
+using ::pluto::layout_stride;
+using ::pluto::default_accessor;
+using ::pluto::extents;
+using ::pluto::dextents;
+using ::pluto::dims;
+using ::pluto::mdspan;
+} // namespace atlas
