--- src/atlas/util/mdspan.h
@@ -1,15 +1,0 @@
-#pragma once
-
-#include "pluto/mdspan.h"
-
-namespace atlas {
-using ::pluto::dynamic_extent;
-using ::pluto::layout_left;
-using ::pluto::layout_right;
-using ::pluto::layout_stride;
-using ::pluto::default_accessor;
-using ::pluto::extents;
-using ::pluto::dextents;
-using ::pluto::dims;
-using ::pluto::mdspan;
-} // namespace atlas
